+++ Classes/UVStyleSheet.m
@@ -16,6 +16,7 @@ static UVStyleSheet *instance;
     if (instance == nil) {
         instance = [UVStyleSheet new];
         instance.loadingViewBackgroundColor = [UIColor colorWithRed:0.902f green:0.902f blue:0.902f alpha:1.0f];
+        instance.preferredStatusBarStyle = UIStatusBarStyleDefault;
     }
     return instance;
 }